+++ prefect_sqlalchemy/__init__.py
@@ -1,4 +1,10 @@
 from . import _version
-from .credentials import DatabaseCredentials, AsyncDriver, SyncDriver  # noqa
+from .credentials import (  # noqa
+    DatabaseCredentials,
+    ConnectionComponents,
+    AsyncDriver,
+    SyncDriver,
+)
+from .database import SqlAlchemyConnector  # noqa
 
 __version__ = _version.get_versions()["version"]